[PATCH v41] Add support of event triggers on authenticated login
The patch introduces trigger on login event, allowing to fire some actions
right on the user connection. This can be useful for  logging or connection
check purposes as well as for some personalization of environment. Usage
details are described in the documentation included, but shortly usage is
the same as for other triggers: create function returning event_trigger and
then create event trigger on login event.